adding files

+++ CommonCore/M0/libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: Gabmiran <Gabmiran@student.42Porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/17 11:54:08 by Gabmiran          #+#    #+#             */
-/*   Updated: 2025/04/17 14:59:47 by Gabmiran         ###   ########.fr       */
+/*   Updated: 2025/05/16 23:31:05 by Gabmiran         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	index;
